Add Phase 2 pipeline, RNN model, AGSS fixes, and visualizations
- agss.py: fix float64 dtype bug, add adaptive eps fallback for high-dim data,
  support min_samples=2 for small minority sets (German dataset)
- models.py: add FraudRNN alongside FraudLSTM
- pipeline_german.py: rewritten to use local CSV (9 features + Risk column),
  best config (RNN, hidden=32, epochs=200, lr=5e-4, threshold=0.8),
  reports F1 for good-credit class matching paper's metric convention
- tune_german.py: hyperparameter grid search with full threshold sweep (0.2–0.8)
- visualize.py: add Phase 2 German bar chart and F1 heatmap

+++ src/agss.py
@@ -124,7 +124,7 @@ class AGSS:
         X_syn = np.array(synthetic)
         y_syn = np.ones(len(X_syn), dtype=y.dtype)
 
-        X_out = np.vstack([X, X_syn])
+        X_out = np.vstack([X, X_syn]).astype(X.dtype)
         y_out = np.concatenate([y, y_syn])
         return X_out, y_out
